+++ src/main/java/io/aero/v2/query/InsertRowQuery.java
@@ -63,6 +63,6 @@ public class InsertRowQuery {
     }
 
     public List<Map<String, String>> getResult() throws SQLException {
-        return new GetRowQuery().setTable(table).setRowId(Long.valueOf(generatedKey).intValue()).generate().execute().getResult();
+        return new GetRowQuery().setTable(table).setRowId(Long.valueOf(generatedKey)).generate().execute().getResult();
     }
 }
